pyproject.toml
[project]
name = "inventory-management"
version = "0.1.0"
description = "\"A system to control inventory\""
authors = [
    {name = "Gabriel Leal Gonçalves",email = "gabrielleal96.gl@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.0,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "alembic (>=1.16.3,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-decouple (>=3.8,<4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.2"
black = "^25.1.0"
mypy = "^1.16.1"
pytest = "^8.4.1"
pytest-mock = "^3.14.1"
pytest-cov = "^6.2.1"
pytest-asyncio = "^1.0.0"

